Time: 8 ms (62.51%), Space: 15.5 MB (58.43%) - LeetHub

+++ 0150-evaluate-reverse-polish-notation/0150-evaluate-reverse-polish-notation.cpp
@@ -1,9 +1,9 @@
 class Solution {
 public:
     int evalRPN(vector<string>& tokens) {
-        stack<long> st;
-        long temp1;
-        long temp2;
+        stack<int> st;
+        int temp1;
+        int temp2;
         for(int i=0; i<tokens.size(); i++){
             if(tokens[i]=="+"){
                 temp1=st.top();
@@ -11,29 +11,25 @@ public:
                 temp2=st.top();
                 st.pop();
                 st.push(temp1+temp2);
-            }
-            else if(tokens[i]=="-"){
+            }else if(tokens[i]=="-"){
                 temp1=st.top();
                 st.pop();
                 temp2=st.top();
                 st.pop();
                 st.push(temp2-temp1);
-            }
-            else if(tokens[i]=="*"){
+            }else if(tokens[i]=="*"){
                 temp1=st.top();
                 st.pop();
                 temp2=st.top();
                 st.pop();
                 st.push(temp1*temp2);
-            }
-            else if(tokens[i]=="/"){
+            }else if(tokens[i]=="/"){
                 temp1=st.top();
                 st.pop();
                 temp2=st.top();
                 st.pop();
                 st.push(temp2/temp1);
-            }
-            else{
+            }else{
                 st.push(stoi(tokens[i]));
             }
         }
